Detect and Remove Loop in a Linked List
Write a function detectAndRemoveLoop() that checks whether a given Linked List 
contains loop and if loop is present then removes the loop and returns true. If the list doesn’t 
contain loop then it returns false. Below diagram shows a linked list with a loop. 
detectAndRemoveLoop() must change the below list to 1->2->3->4->5->NULL.